*moved sources to 'lib' directory
