+++ app/actors/tax/taxcloud/gotaxcloud/authorized.go
@@ -1,4 +1,4 @@
-package taxcloud
+package gotaxcloud
 
 import (
 	"encoding/json"
