+++ app/actors/tax/taxcloud/gotaxcloud/types.go
@@ -1,4 +1,4 @@
-package taxcloud
+package gotaxcloud
 
 import (
 	"errors"
